Refactor: Centralized job fetching with shared parser and execution time benchmarking
