+++ Scripts/__init__.py
@@ -1,0 +1,5 @@
+import sys
+import os
+
+# Add parent directory to the path
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
